+++ tests/e2e/tests/helpers/learncard.helpers.ts
@@ -1,13 +1,21 @@
 import { readFile } from 'fs/promises';
 
-import { initLearnCard } from '@learncard/init';
+import { AddPlugin } from '@learncard/core';
+import { initLearnCard, NetworkLearnCardFromSeed } from '@learncard/init';
+import { getSimpleSigningPlugin, SimpleSigningPlugin } from '@learncard/simple-signing-plugin';
 
 const didkit = readFile(
     require.resolve('@learncard/didkit-plugin/dist/didkit/didkit_wasm_bg.wasm')
 );
 
-export const getLearnCard = (seed = 'a'.repeat(64), managedDid?: string, debug = false) => {
-    return initLearnCard({
+export type LearnCard = AddPlugin<NetworkLearnCardFromSeed['returnValue'], SimpleSigningPlugin>;
+
+export const getLearnCard = async (
+    seed = 'a'.repeat(64),
+    managedDid?: string,
+    debug = false
+): Promise<LearnCard> => {
+    const learnCard = await initLearnCard({
         seed,
         didkit,
         network: 'http://localhost:4000/trpc',
@@ -15,6 +23,10 @@ export const getLearnCard = (seed = 'a'.repeat(64), managedDid?: string, debug =
         ...(managedDid && { didWeb: managedDid }),
         ...(debug && { debug: console.log }),
     });
+
+    return learnCard.addPlugin(
+        await getSimpleSigningPlugin(learnCard, 'http://localhost:4200/trpc')
+    ) as any;
 };
 
 export const USERS = {