+++ packages/learn-card-network/cloud-client/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @learncard/network-brain-client
 
+## 1.3.12
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/learn-cloud-service@2.2.15
+
+## 1.3.11
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/learn-cloud-service@2.2.14
+
+## 1.3.10
+
+### Patch Changes
+
+-   Updated dependencies [[`60090c28bbd2af3b659e0e6a6f648cb1ddf5efaa`](https://github.com/learningeconomy/LearnCard/commit/60090c28bbd2af3b659e0e6a6f648cb1ddf5efaa)]:
+    -   @learncard/learn-cloud-service@2.2.13
+
+## 1.3.9
+
+### Patch Changes
+
+-   Updated dependencies [[`b9beeb4bd8abfecb9811b1d80a0b7c9aa6386114`](https://github.com/learningeconomy/LearnCard/commit/b9beeb4bd8abfecb9811b1d80a0b7c9aa6386114)]:
+    -   @learncard/learn-cloud-service@2.2.12
+
 ## 1.3.8
 
 ### Patch Changes